code polishes and rename
